requirements.txt
transformers>=4.30.0
torch>=2.0.0
biopython>=1.81
numpy>=1.24.0
tqdm>=4.65.0
scikit-learn>=1.3.0
